+++ tests/test_cli.py
@@ -1,5 +1,3 @@
-from pathlib import Path
-
 from click.testing import CliRunner
 
 from hubspot_tech_writing.cli import cli
@@ -19,26 +17,40 @@ def test_version():
     assert result.exit_code == 0
 
 
-def test_convert_file_to_stdout():
-    infile = Path(__file__).parent / "data" / "hubspot-blog-post-original.md"
+def test_convert_file_to_stdout(markdownfile):
     runner = CliRunner()
 
     result = runner.invoke(
         cli,
-        args=f"--debug convert '{infile}'",
+        args=f"--debug convert '{markdownfile}'",
         catch_exceptions=False,
     )
     assert result.exit_code == 0
 
 
-def test_convert_file_to_file(tmp_path):
-    infile = Path(__file__).parent / "data" / "hubspot-blog-post-original.md"
+def test_convert_file_to_file(tmp_path, markdownfile):
     outfile = tmp_path / "converted.html"
     runner = CliRunner()
 
     result = runner.invoke(
         cli,
-        args=f"convert '{infile}' '{outfile}'",
+        args=f"convert '{markdownfile}' '{outfile}'",
         catch_exceptions=False,
     )
     assert result.exit_code == 0
+
+
+def test_linkcheck_broken(caplog, markdownfile_minimal_broken_links):
+    runner = CliRunner()
+
+    result = runner.invoke(
+        cli,
+        args=f"--debug linkcheck '{markdownfile_minimal_broken_links}'",
+        catch_exceptions=False,
+    )
+    assert result.exit_code == 22
+
+    assert "Total files checked: 1" in result.output
+    assert "Broken image: images/bar.png" in caplog.text
+    assert "[✖] images/bar.png" in result.output
+    assert "[✖] https://foo.example.org/" in result.output