+++ CHANGES.md
@@ -7,6 +7,7 @@
 - CI: Add GHA workflow recipes
 - Add initial converter implementation
 - HTML: Add addons for custom code blocks, and header links
+- Add linkcheck subsystem
 
 
 ## 2023/09/xx 0.0.0